+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- `ExceptionUtils` class to unify exceptions format, by @HardNorth
 ### Changed
 - `MarkdownUtils`, `TemplateConfiguration` and `TemplateProcessing` was moved to `utils.formatting` package, by @HardNorth